Adding the ability to attach normal listeners for scroll operations.
Namely listeners for AbsListView.OnScrollListener, RecyclerView.OnScrollListener, or ObservableScrollView.OnScrollChangedListener.
This allows for users of this library to be able to still receive normal
view callbacks in addition to the onScrollUp and onScrollDown callbacks implemented
by this library. This does not change any of the existing interfaces to this library and
is backwards compatible. New attach methods are provided if the user wants to be notified
of the normal view callbacks.